# openapi.yaml
openapi: 3.0.0
info:
  title: Java Faktury Project API
  description: "API for project about generating invoices using Java, Spring Boot,\
    \ MongoDB and React"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://faktury-backend.herokuapp.com/api
  description: API Server
security:
- bearerAuth: []
...
paths:
...
  /invoice:
    get:
      tags:
      - invoice
      summary: |
        search for existing invoices - retrieves all if searchstr blank (length = 0 or just a space)
      operationId: getInvoices
      parameters:
      - name: title
        in: query
        description: Title of invoice to filter with
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quantity
        in: query
        description: The number of invoices to return. Fetches all qualifying if not
          specified
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: before
        in: query
        description: Date to filter out invoices before specified day
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: after
        in: query
        description: Date to filter out invoices after specified day
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: contractor
        in: query
        description: ID of a contractor to get invoices from
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "500":
          description: internal server error
    post:
      tags:
      - invoice
      summary: create new invoice
      operationId: newInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        "201":
          description: invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_201'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "500":
          description: internal server error
  /invoice/pdf:
    get:
      tags:
      - invoice
      summary: get filtered invoices summary in pdf
      operationId: getInvoicesPdf
      parameters:
      - name: title
        in: query
        description: Title of invoice to filter with
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quantity
        in: query
        description: The number of invoices to return. Fetches all qualifying if not
          specified
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: before
        in: query
        description: Date to filter out invoices before specified day
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: after
        in: query
        description: Date to filter out invoices after specified day
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: contractor
        in: query
        description: ID of a contractor to get invoices from
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "500":
          description: internal server error
  /invoice/{id}:
    get:
      tags:
      - invoice
      summary: get specific invoice
      operationId: getInvoice
      parameters:
      - name: id
        in: path
        description: ID of the invoice to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        "400":
          description: wrong request
        "404":
          description: no such invoice
        "401":
          description: unauthorized
        "500":
          description: internal server error
    put:
      tags:
      - invoice
      summary: update invoice
      operationId: updateInvoice
      parameters:
      - name: id
        in: path
        description: ID of the invoice to update
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such contractor
        "500":
          description: internal server error
    delete:
      tags:
      - invoice
      summary: delete specified invoice
      operationId: deleteInvoice
      parameters:
      - name: id
        in: path
        description: ID of the invoice to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such invoice
        "500":
          description: internal server error
  /invoice/{id}/pdf:
    get:
      tags:
      - invoice
      summary: get specific invoice in pdf
      operationId: getInvoicePdf
      parameters:
      - name: id
        in: path
        description: ID of the invoice to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such invoice
        "500":
          description: internal server error
  /contractor:
    get:
      tags:
      - contractor
      summary: |
        search for existing contractrors - retrieves all if searchstr blank (length = 0 or just a space)
      operationId: getContractors
      parameters:
      - name: searchstr
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: Media
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contractor'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "500":
          description: internal server error
    post:
      tags:
      - contractor
      summary: add new contractor to database
      operationId: newContractor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contractor'
        required: true
      responses:
        "201":
          description: contractor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_201'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "409":
          description: contractor already exists
        "500":
          description: internal server error
  /contractor/{id}:
    get:
      tags:
      - contractor
      summary: get contractor data with ID
      operationId: getContractorWithId
      parameters:
      - name: id
        in: path
        description: ID of the contractor to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contractor'
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: contractor not found
        "500":
          description: internal server error
    put:
      tags:
      - contractor
      summary: update contractor data
      operationId: updateContractor
      parameters:
      - name: id
        in: path
        description: ID of the contractor to update
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contractor'
        required: true
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such contractor
        "500":
          description: internal server error
    delete:
      tags:
      - contractor
      summary: delete specified contractor
      operationId: deleteContractor
      parameters:
      - name: id
        in: path
        description: ID of the contractor to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: success
        "400":
          description: wrong request
        "401":
          description: unauthorized
        "404":
          description: no such contractor
        "500":
          description: internal server error
...
components:
  schemas:
...
    InvoiceItem:
      allOf:
      - $ref: '#/components/schemas/Product'
      - required:
        - quantity
        type: object
        properties:
          quantity:
            type: number
            example: 2.5
    Invoice:
      required:
      - contractor
      - created
      - dueDate
      - issueDate
      - issuePlace
      - products
      - title
      type: object
      properties:
        id:
          type: string
        contractor:
          $ref: '#/components/schemas/Contractor'
        title:
          type: string
          example: 5324/04/2021
        products:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItem'
        created:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        edited:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        issuePlace:
          type: string
          example: Wrocław
        issueDate:
          type: string
          format: date
          example: 2016-08-29
        sellDate:
          type: string
          format: date
          example: 2016-08-29
        paymentDate:
          type: string
          format: date
          example: 2016-08-30
        paymentType:
          type: string
          example: Płatność gotówką
    Product:
      required:
      - currency
      - name
      - price
      - tax
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: straws
        price:
          type: number
          example: 35.9
        tax:
          type: integer
          example: 23
        currency:
          type: string
          example: PLN
        unit:
          type: string
          example: szt.
    Address:
      type: object
      properties:
        city:
          type: string
          example: Warszawa
        street:
          type: string
          example: Al. Jerozolimskie 179
        postalCode:
          type: string
          example: 02-222
    Contractor:
      required:
      - address
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: Media Saturn Online Sp. z o.o.
        address:
          $ref: '#/components/schemas/Address'
        nip:
          type: string
          example: "1234567890"
        phoneNumber:
          type: string
          example: 123 456 789
        email:
          type: string
          format: email
          example: email@email.com
        bankName:
          type: string
          example: ALIOR BANK SA
        bankAccountNumber:
          type: string
          example: PL12 3456 0000 0000 0000 0000 7890
    inline_response_201:
      type: object
      properties:
        id:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT